requirements.txt
pupa
pyyaml
python-dateutil
unidecode
pytz